+++ sphinx/highlighting.py
@@ -229,7 +229,7 @@ class PygmentsBridge:
             # MEMO: this is done to escape Unicode chars with non-Unicode engines
             return texescape.hlescape(hlsource, self.latex_engine)
 
-    def get_stylesheet(self, selectors: Optional[List[int]] = None) -> str:
+    def get_stylesheet(self, selectors: list[int] | None = None) -> str:
         formatter = self.get_formatter()
         if self.dest == 'html':
             if selectors: